delete conversion controller
